requirements.txt
fastapi
uvicorn
sqlalchemy
jinja2
python-slugify
pydantic
